fix: add default parameter

+++ app/db/scripts/procedures/drop_table.sql
@@ -1,7 +1,9 @@
-CREATE OR REPLACE PROCEDURE drop_table(table_name TEXT)
+CREATE OR REPLACE PROCEDURE drop_table(table_name TEXT, explicity BOOLEAN DEFAULT TRUE)
 AS $$
 BEGIN
-    EXECUTE FORMAT('DELETE FROM public.%I', table_name);
+    IF explicity THEN
+        EXECUTE FORMAT('DELETE FROM public.%I', table_name);
+    END IF;
     EXECUTE FORMAT('DROP TABLE IF EXISTS public.%I CASCADE', table_name);
     RAISE NOTICE 'Table % has been dropped successfully.', table_name;
 END;
